+++ plone/app/event/browser/event_view.py
@@ -1,9 +1,4 @@
-from AccessControl import getSecurityManager
-from Acquisition import aq_parent
-from Products.CMFCore import permissions
 from Products.Five.browser import BrowserView
-from Products.statusmessages.interfaces import IStatusMessage
-from plone.app.event import messageFactory as _
 from plone.event.interfaces import IEventAccessor
 from plone.event.interfaces import IOccurrence
 
@@ -34,21 +29,7 @@ class EventView(BrowserView):
 
     def __call__(self):
         if IOccurrence.providedBy(self.context):
-            # If the context is a Occurrence, and the user has edit rights,
-            # tell the user to go onle level up to be able to edit the event
-            # via a IStatusMessage.
-            # Actually, a HTML link would be better, but that's not supported
-            # by the globalstatusmessage template at the moment
-            # (https://github.com/plone/Products.CMFPlone/issues/206).
-            event = aq_parent(self.context)
-            sm = getSecurityManager()
-            can_edit = sm.checkPermission(
-                permissions.ModifyPortalContent, event)
-            if can_edit:
-                msg = _(
-                    'part_of_recurring_event',
-                    default=u'This event is part of a recurring Event. '
-                            u'To edit the original event, go one level up.'
-                )
-                IStatusMessage(self.request).addStatusMessage(msg, 'info')
+            # The transient Occurrence objects cannot be edited. disable the
+            # edit border for them.
+            self.request.set('disable_border', True)
         return self.index()  # render me.