+++ plone/app/event/recurrence.py
@@ -126,7 +126,6 @@ class Occurrence(SimpleItem):
     recurrecne set.
     """
     implements(IOccurrence)
-    _Copy_or_Move_Permission = []
 
     def __init__(self, id, start, end):
         self.id = id